complete: restructure entire app with consistent organization
