+++ package.json
@@ -15,8 +15,7 @@
     "@radix-ui/react-progress": "1.0.3",
     "@radix-ui/react-slot": "1.0.2",
     "@radix-ui/react-tabs": "1.0.4",
-    "lucide-react": "0.303.0",
-    "framer-motion": "^10.17.9"
+    "lucide-react": "0.303.0"
   },
   "devDependencies": {
     "typescript": "5.3.3",
